Finish working with the new functionality - Import/Export user data, fixing the History part, fixing the problems with the RouteChanges, fixing some bugs.

+++ sofbus24/src/main/java/bg/znestorov/sofbus24/utils/activity/SearchEditText.java
@@ -96,8 +96,8 @@ public class SearchEditText extends EditText {
 
                 if (!bounds.contains(actionX, actionY)) {
                     /** Gives the +20 area for tapping. */
-                    x = (int) (actionX - extraTapArea);
-                    y = (int) (actionY - extraTapArea);
+                    x = actionX - extraTapArea;
+                    y = actionY - extraTapArea;
 
                     if (x <= 0)
                         x = actionX;
@@ -134,8 +134,8 @@ public class SearchEditText extends EditText {
                  * BOUND. - this process help to increase the tappable area of
                  * the rectangle.
                  */
-                x = (int) (actionX + extraTapArea);
-                y = (int) (actionY - extraTapArea);
+                x = actionX + extraTapArea;
+                y = actionY - extraTapArea;
 
                 /**
                  * Since this is right drawable subtract the value of x from the
